dashboard access problem solved

+++ admin_app/routes/web.php
@@ -24,7 +24,7 @@ Route::middleware(['auth'])->group(function () {
         Route::get('users', [UserController::class, 'index'])->name('users.index');
     });
 
-    Route::middleware(['auth','permission:dashboard.view'])->group(function () {
+    Route::middleware(['auth', 'permission:dashboard.view'])->group(function () {
         Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
     });
 });
@@ -44,17 +44,18 @@ Route::get('foods', function () {
 
 Route::prefix('admin')->middleware(['auth'])->group(function () {
 
-    Route::middleware('permission:role.view')->get('roles', [RoleController::class, 'index']);
+    Route::middleware('permission:role.view')->get('roles', [RoleController::class, 'index'])->name('roles.index');
     Route::middleware('permission:role.edit')->get('roles/{role}/permissions', [RoleController::class, 'editPermissions'])->name('roles.permissions.edit');
     Route::middleware('permission:role.edit')->post('roles/{role}/permissions', [RoleController::class, 'updatePermissions'])->name('roles.permissions.update');
-    Route::middleware('permission:role.create')->post('roles', [RoleController::class, 'store']);
-    Route::middleware('permission:role.edit')->put('roles/{role}', [RoleController::class, 'update']);
-    Route::middleware('permission:role.delete')->delete('roles/{role}', [RoleController::class, 'destroy']);
+    Route::middleware('permission:role.create')->post('roles', [RoleController::class, 'store'])->name('roles.store');
+    Route::middleware('permission:role.edit')->put('roles/{role}', [RoleController::class, 'update'])->name('roles.update');
+    Route::middleware('permission:role.delete')->delete('roles/{role}', [RoleController::class, 'destroy'])->name('roles.destroy');
 
-    Route::middleware('permission:permission.view')->get('permissions', [PermissionController::class, 'index']);
-    Route::middleware('permission:permission.create')->post('permissions', [PermissionController::class, 'store']);
-    Route::middleware('permission:permission.edit')->put('permissions/{permission}', [PermissionController::class, 'update']);
-    Route::middleware('permission:permission.delete')->delete('permissions/{permission}', [PermissionController::class, 'destroy']);
+
+    Route::middleware('permission:permission.view')->get('permissions', [PermissionController::class, 'index'])->name('permissions.index');
+    Route::middleware('permission:permission.create')->post('permissions', [PermissionController::class, 'store'])->name('permissions.store');
+    Route::middleware('permission:permission.edit')->put('permissions/{permission}', [PermissionController::class, 'update'])->name('permissions.update');
+    Route::middleware('permission:permission.delete')->delete('permissions/{permission}', [PermissionController::class, 'destroy'])->name('permissions.destroy');
 });
 
 require __DIR__ . '/auth.php';
